+++ src/main/java/com/mycompany/myapp/service/impl/UserContentPraiseServiceImpl.java
@@ -1,6 +1,9 @@
 package com.mycompany.myapp.service.impl;
 
+
+import com.mycompany.myapp.domain.ContentInfo;
 import com.mycompany.myapp.domain.UserAccount;
+import com.mycompany.myapp.repository.ContentInfoRepository;
 import com.mycompany.myapp.repository.UserAccountRepository;
 import com.mycompany.myapp.service.UserContentPraiseService;
 import com.mycompany.myapp.domain.UserContentPraise;
@@ -13,7 +16,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.security.PublicKey;
 import java.time.ZonedDateTime;
 import java.util.Optional;
 
@@ -28,11 +30,12 @@ public class UserContentPraiseServiceImpl implements UserContentPraiseService {
 
     private final UserContentPraiseRepository userContentPraiseRepository;
     private final UserAccountRepository userAccountRepository;
+    private ContentInfoRepository contentInfoRepository;
 
-    public UserContentPraiseServiceImpl(UserContentPraiseRepository
-                                            userContentPraiseRepository, UserAccountRepository userAccountRepository) {
+    public UserContentPraiseServiceImpl(UserContentPraiseRepository userContentPraiseRepository, UserAccountRepository userAccountRepository, ContentInfoRepository contentInfoRepository) {
         this.userContentPraiseRepository = userContentPraiseRepository;
-        this.userAccountRepository = userAccountRepository;
+        this.userAccountRepository=userAccountRepository;
+        this.contentInfoRepository = contentInfoRepository;
     }
 
     @Override
@@ -56,29 +59,58 @@ public class UserContentPraiseServiceImpl implements UserContentPraiseService {
         return userContentPraiseRepository.findById(id);
     }
 
+
     @Override
-    public void delete(Long id) {
+    public void delete(String login,Long id) throws Exception {
         log.debug("Request to delete UserContentPraise : {}", id);
-        userContentPraiseRepository.deleteById(id);
+        Optional<UserContentPraise> praiseOptional=this.userContentPraiseRepository.findById(id);
+        if (praiseOptional.isPresent()){
+            //2.判断数据对象是否是当前登录用户的
+            UserContentPraise praise=praiseOptional.get();
+            if (praise.getAccount()!=null
+                && praise.getAccount().getLogin()!=null
+                && praise.getAccount().getLogin().equals(login)){
+                userContentPraiseRepository.deleteById(id);
+                if (praise.getContent()!=null) {
+                    ContentInfo contentInfo = praise.getContent();
+                    Long praiseCount = contentInfo.getContentPraiseCount();
+                    if (praiseCount == null) {
+                        praiseCount = 0l;
+                    } else {
+                        if (praiseCount > 1) {
+                            praiseCount -= 1;
+                        } else {
+                            praiseCount = 0l;
+                        }
+                    }
+                    contentInfo.setContentPraiseCount(praiseCount);
+                    this.contentInfoRepository.save(contentInfo);
+                 }
+                }else{
+                    throw new Exception("id为"+id+"的数据不属于当前登录用户"+login);
+                }
+        }else{
+            throw new Exception("未找到id为"+id+"的数据");
+        }
     }
 
     @Override
     public UserContentPraise praise(String login, UserContentPraise praise) {
-        UserAccount userAccount = this.userAccountRepository.findByLogin(login);
-        if (userAccount != null) {
+        praise.setPraiseTime(ZonedDateTime.now());
+        UserAccount userAccount=this.userAccountRepository.findByLogin(login);
+        if(userAccount!=null){
             praise.setAccount(userAccount);
         }
-        praise.setPraiseTime(ZonedDateTime.now());
-
-        UserContentPraise existsPraise = this.userContentPraiseRepository.
-            findByContentIdAndAccountLogin(praise.getContent().getId(), login);
-        if (existsPraise != null) {
+        UserContentPraise existsPraise=this.userContentPraiseRepository
+            .findByContentIdAndAccountLogin(praise.getContent().getId(),login);
+        //如果已经存在点赞数据，对原点赞对象的点赞时间进行更新，并存入数据库
+        if(existsPraise!=null){
             existsPraise.setPraiseTime(ZonedDateTime.now());
-            praise = this.save(existsPraise);
-        } else {
-            praise = this.save(praise);
+        }else {//如果没有点赞数据，表示第一次点赞，直接保存点赞对象
+            praise=this.save(praise);
         }
-
         return praise;
     }
 }
+
+
